move maven and gradle version loader to the modularized generator
