+++ generators/bootstrap-application-server/generator.ts
@@ -35,14 +35,14 @@ import { loadRequiredConfigIntoEntity, prepareEntityPrimaryKeyForTemplates } fro
 import {
   loadRequiredConfigDerivedProperties,
   prepareEntity as prepareEntityServerForTemplates,
-  getPomVersionProperties,
-  getGradleLibsVersionsProperties,
   addEntitiesOtherRelationships,
   hibernateSnakeCase,
   loadServerConfig,
   loadDerivedServerConfig,
   prepareRelationship,
 } from '../server/support/index.js';
+import { getGradleLibsVersionsProperties } from '../gradle/support/index.js';
+import { getPomVersionProperties } from '../maven/support/index.js';
 import { prepareField as prepareFieldForLiquibaseTemplates } from '../liquibase/support/index.js';
 import { dockerPlaceholderGenerator, getDockerfileContainers } from '../docker/utils.js';
 import { GRADLE_VERSION } from '../gradle/constants.js';
