+++ generators/gradle/support/index.ts
@@ -17,6 +17,4 @@
  * limitations under the License.
  */
 
-export * from './dependabot-maven.js';
-export { default as PomStorage } from './pom-store.js';
-export * from './pom-store.js';
+export * from './dependabot-gradle.js';
