+++ generators/maven/support/dependabot-maven.ts
